# apispec.yaml
openapi: 3.0.0
info:
  title: docsis-pnm API
  version: 2.1.0
servers: 
  - url: /
paths:
...
  /cmts:
    get:
      tags:
        - CMTS
      responses:
        200:
          description: List all CMT systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cmts'
    post:
      tags: 
        - CMTS
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cmts'
      responses:
        201:
          $ref: '#/components/responses/cmtsResponse'
...
  /modems/{modemId}:
    get:
      tags:
        - Modem
      parameters:
        - $ref: '#/components/parameters/modemId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Modem'
        404:
          description: Not found
    patch:
      tags:
        - Modem
        - Not implemented yet
      parameters:
        - $ref: '#/components/parameters/modemId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Modem'
      responses:
        200:
          description: Update successful
        404:
          description: Not found
        409:
          $ref: '#/components/responses/validationError'
    delete:
      tags:
        - Modem
        - Not implemented yet
      parameters:
        - $ref: '#/components/parameters/modemId'
      description: Delete the modem from the database. History data
        IS NOT DELETED because this is a very I/O expensive operation.
        History data deletion relies on database partitions.
        If this modem is active on any CMTS it will be automatically created
        at the next CMTS poll run.
      responses:
        200:
          description: OK
...
  /modems/{modemId}/downstreamHistory/latest:
    get:
      tags:
        - Modem
      parameters:
        - $ref: '#/components/parameters/modemId'
      description: latest downstream performance data as returned from modem
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownstreamHistory'
...
  /modems/{modemId}/upstreamHistoryFromCMTS/latest:
    get:
      tags:
        - Modem
      parameters:
        - $ref: '#/components/parameters/modemId'
      description: latest upstream performance data as returned from CMTS
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpstreamHistory'
...
  /modems/{modemId}/upstreamHistoryFromModem/latest:
    get:
      tags: 
        - Modem
      parameters: 
        - $ref: '#/components/parameters/modemId'
      description: latest upstream performance data as returned from modem
      responses:
        200:
          headers:
            X-Count:
              $ref: '#/components/headers/X-Count'
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpstreamHistoryModem'
...
components:
  headers:
    X-Count:
      description: Number of returned items
      schema:
        type: integer
  responses:
    validationError:
      description: ValidationError
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    cmtsResponse:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Cmts'
  parameters:
...
    modemId:
      name: modemId
      in: path
      required: true
      schema:
        type: string
      description: Modem identifier. Can be the MAC address, database row ID
        or IP address of the modem
...
  schemas:
...
    Error:
      type: object
      readOnly: true
      properties:
        message:
          type: string
        field_errors:
          type: object
          description: Fields with error message
          example: {'hostname': 'invalid hostname'}
    Cmts:
      type: object
      required:
        - hostname
      properties:
        cmts_id:
          type: integer
          format: uint32
          example: 1
          readOnly: true
        hostname:
          description: Hostname or IP address
          type: string
          example: cmts-rosenheim.example
        snmp_community:
          type: string
          example: public
          nullable: true
          description: Override default SNMP community from config file.
            Returns the overridden community string if set. Set to `null` to
            use the default community.
        snmp_community_modem:
          type: string
          example: public
          nullable: true
          description: Override the default SNMP community for modems.
            Returns the overridden community string if set. Set to `null` to
            use the default community.
        disabled:
          type: boolean
          example: false
          nullable: false
          description: Set to `true` to disable polling of that CMTS
        poll_interval:
          type: integer
          example: 60
          nullable: false
          description: Poll interval of the CMTS. Defaults to 60 if < 60
...
    Modem:
      properties:
        modem_id:
          type: integer
          format: uint64
          readOnly: true
          example: 1
        cmts_id:
          type: integer
          format: int32
          readOnly: true
          example: 1
          description: The ID of the CMTS the modem was last seen on
        mac:
          type: string
          format: macaddr
          readOnly: true
          description: MAC address in Unix format (00:00:00:00:00:00)
          example: 00:00:00:00:00:00
        sysdescr:
          type: string
          readOnly: true
          description: Sysdescr as returned via SNMP
          example: 'FRITZ!Box Cable <<HW_REV: 233.8; VENDOR: AVM GmbH; BOOTR: 1.3465; SW_REV: 161.07.19-77042; MODEL: 6591>>'
        ipaddr:
          type: string
          readOnly: true
          description: IP address (IPv4 or IPv6) of the cable modem part of the modem
        snmp_index:
          type: integer
          format: int32
          description: SNMP index on the CMTS for that modem
          readOnly: true
        docsis_version:
          type: string
          enum:
            - 'docs_10'
            - 'docs_11'
            - 'docs_20'
            - 'docs_21'
            - 'docs_31'
            - 'unknown'
          readOnly: true
          description: DOCSIS version this modem is running on
...
    UpstreamHistoryEntry:
      properties:
        id:
          type: integer
          format: int32
          description: Upstream ID
        pwr:
          type: integer
          format: int32
          description: Power in tenth dBmV
          example: -10
        snr:
          type: integer
          format: int32
          description: Signal to noise ratio (SNR) in tenth dB (you have to divide the value by ten)
          example: 299
        mrefl:
          type: integer
          format: int32
          example: -30
          description: Microreflections in -dBc
        unerr:
          type: integer
          format: uint64
          description: Number of received unerrored codewords
        corr:
          type: integer
          format: uint64
          description: Number of received corrected codewords
        err:
          type: integer
          format: uint64
          description: Number of erroneous codewords
    UpstreamHistory:
      properties:
        ts:
          type: integer
          format: int64
          description: Unix timestamp in full seconds
        us:
          type: array
          items:
            $ref: '#/components/schemas/UpstreamHistoryEntry'
    DownstreamChannel:
      properties:
        freq:
          type: integer
          format: int32
          description: Downstream channel frequency in Hertz
          example: 800000000
        pwr:
          type: integer
          format: int32
          description: Receive power in tenth dbmV (you have to divide the value by ten)
          example: -12
        snr:
          type: integer
          format: int32
          description: Signal to noise ratio (SNR) in tenth dB (you have to divide the value by ten)
          example: 401
        mrefl:
          type: integer
          format: int32
          example: -30
          description: Microreflections in -dBc
        unerr:
          type: integer
          format: uint64
          description: Number of received unerrored codewords
        corr:
          type: integer
          format: uint64
          description: Number of received corrected codewords
        err:
          type: integer
          format: uint64
          description: Number of erroneous codewords
        mod:
          type: integer
          format: int32
          description: Modulation of the channel
...
    DownstreamHistory:
      properties:
        ts:
          type: integer
          format: int64
          description: Unix timestamp in full seconds
        ds:
          type: array
          items:
            $ref: '#/components/schemas/DownstreamChannel'
    UpstreamHistoryModem:
      properties:
        ts:
          type: integer
          format: int54
          description: Unix timestamp in full seconds
        us:
          type: array
          items:
            $ref: '#/components/schemas/ModemUpstream'
    ModemUpstream:
      properties:
        freq:
          type: integer
          format: int32
          description: Upstream Frequency in Hz
        timing_offset:
          type: integer
          format: int32
          description: Timing offset
        tx_power:
          type: integer
          format: int32
          description: Transmit power of the channel in tenth dBmV
...